requirements.txt
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.26.1
pdfkit==0.6.1
PyPDF2==1.26.0
beautifulsoup4==4.10.0
html2text==2020.1.16